fix: full width layout in Search and Notifications
- remove unused AppNavbar
- merge AppSearch and Search
